LayoutTests/media/track/track-css-all-cues-expected.txt

Test that style to all cues is applied correctly.
EVENT(canplaythrough)
EVENT(seeked)
EXPECTED (getComputedStyle(textTrackDisplayElement(video).firstChild).color == 'rgb(255, 0, 0)') OK
END OF TEST
